pyproject.toml
[tool.poetry]
name = "bullish"
version = "0.1.0"
description = ""
authors = ["aan <andoludovic.andriamamonjy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
bearishpy = "^0.18.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
mypy = "^1.8.0"
nox = "^2023.4.22"
ruff = "^0.3.0"
pre-commit = "^3.6.2"
pandas-stubs = "^2.2.0.240218"
types-simplejson = "^3.19.0.20240218"
types-beautifulsoup4 = "^4.12.0.20240229"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
plugins = "pydantic.mypy"
packages = "bullish"


follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
show_absolute_path = true
# for strict mypy: (this is the tricky one :-))
disallow_untyped_defs = true
strict = true
#disable_error_code = "import-untyped"

exclude = [
    "test_*",
    "conftest.py",
]

[[tool.mypy.overrides]]
module = [
    "plotly.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 96
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "test_*",
    "conftest.py",
    "*env.py"
]


[tool.ruff.lint]
select = ["E", "W", "F", "B006","RUF","UP","SIM","FURB","PIE","PERF","C4","B","PL","S","C901","N","ANN", "ERA001", "PD"]
ignore = []
extend-ignore = ["ANN101","ANN102","PLR6301","E203","W291","ANN401"]

fixable = ["ALL"]
unfixable = []
